pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipy-awc"
version = "0.9.0"
description = "PiPy Automatic-Water Controller: A dynamically-configurable & automated aquarium water level controller for the Raspberry Pi using Python. Intended to run using mechanical voltage relays (i.e. Waveshare Relay Hat) and optical contact liquid sensors (i.e. from DFRobot/CQRObot)"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Home Automation",
    "Topic :: System :: Hardware",
]
keywords = ["Raspberry Pi", "Aquarium", "Automation", "AWC", "RPi"]
authors = [{ name = "Neil Kuehnle" }]
license = { text = "GPLv2" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apprise>=1.7.4",
    "pyaml>=21.10",
    "schedule>=1.1.0",
    "gpiozero>=1.6.2",
    "imap-tools>=1.5.0",
    "python-dateutil>=2.90",
]

[project.urls]
"Homepage" = "https://github.com/nkuehnle/PiPyAWC"
"Bug Tracker" = "https://github.com/nkuehnle/PiPyAWC/issues"
